chore: rename and move api_events and worker_events
